+++ Base-Project-react/src/routes.tsx
@@ -25,11 +25,11 @@ import ListVariant from "./components/admin/Products/ListVariant";
 import Home from "./components/component/layout/home";
 import CartProduct from "./components/client/Cart/cart";
 import Statistics from "./components/admin/statistics/statistics";
-// import PaymentOnline from "./components/client/payment/paymentOnline";
-// import ConfirmPayment from "./components/client/payment/confirmPayment";
+import PaymentOnline from "./components/client/payment/paymentOnline";
+import ConfirmPayment from "./components/client/payment/confirmPayment";
 import Bill from "./components/client/bill/Bill";
-import Account from "./components/client/account";
-import AccountUpdate from "./components/client/accountUpdate";
+import Account from "./components/client/account/account";
+import AccountUpdate from "./components/client/account/accountUpdate";
 
 
 
@@ -70,15 +70,15 @@ export const router = createBrowserRouter([
                 path: 'cart',
                 element: <CartProduct />
             },
-            // {
-            //     path: 'payment',
-            //     element: <PaymentOnline />
-            // },
+            {
+                path: 'payment',
+                element: <PaymentOnline />
+            },
 
-            // {
-            //     path: "confirm",
-            //     element: <ConfirmPayment />,
-            // },
+            {
+                path: "confirm",
+                element: <ConfirmPayment />,
+            },
             {
                 path: "bill",
                 element: <Bill />
